functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
google-cloud-secret-manager>=2.0.0
httpx>=0.25.0
pyyaml>=6.0
